move crowsnest.py to top level
Having some strange git issues. This is my trying to debug it. When I
created the file in the 02_crowsnest/ directory, git doesn't seem to see
that it's been added or changed. I can't understand why. I made it using
the author's bin/new command, but that doesn't seem to explain it as if
I do that for a test file git sees the new file.